+++ src/bpp/models/dyscyplina_naukowa.py
@@ -35,28 +35,32 @@ class Autor_DyscyplinaManager(models.Manager):
             elem.widoczna = False
             elem.save()
 
-        for elem in self.all().values("dyscyplina").distinct():
-            elem = Dyscyplina_Naukowa.objects.get(pk=elem['dyscyplina'])
-            elem.widoczna = True
-            elem.save()
+        for attr in "dyscyplina_naukowa", "subdyscyplina_naukowa":
+            for elem in self.all().values(attr).distinct():
+                if elem[attr] is None:
+                    continue
 
-            i = elem.dyscyplina_nadrzedna
-            while i is not None:
-                if not i.widoczna:
-                    i.widoczna = True
-                    i.save()
-                i = i.dyscyplina_nadrzedna
+                elem = Dyscyplina_Naukowa.objects.get(pk=elem[attr])
+                elem.widoczna = True
+                elem.save()
+
+                i = elem.dyscyplina_nadrzedna
+                while i is not None:
+                    if not i.widoczna:
+                        i.widoczna = True
+                        i.save()
+                    i = i.dyscyplina_nadrzedna
 
 
 class Autor_Dyscyplina(models.Model):
     rok = PositiveSmallIntegerField()
     autor = models.ForeignKey("bpp.Autor", CASCADE)
 
-    dyscyplina = models.ForeignKey("bpp.Dyscyplina_Naukowa", CASCADE)
-    procent_dyscypliny = models.DecimalField(max_digits=5, decimal_places=2)
+    dyscyplina_naukowa = models.ForeignKey("bpp.Dyscyplina_Naukowa", models.SET_NULL, related_name="dyscyplina", blank=True, null=True)
+    procent_dyscypliny = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
 
-    subdyscyplina = models.ForeignKey("bpp.Dyscyplina_Naukowa", CASCADE)
-    procent_subdyscypliny = models.DecimalField(max_digits=5, decimal_places=2)
+    subdyscyplina_naukowa = models.ForeignKey("bpp.Dyscyplina_Naukowa", models.SET_NULL, related_name="subdyscyplina", blank=True, null=True)
+    procent_subdyscypliny = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
 
     objects = Autor_DyscyplinaManager()